ajustes de duplicação no sonarCloud
