+++ package.json
@@ -64,7 +64,8 @@
     "collectCoverageFrom": [
       "src/App.js",
       "src/(components|pages)/**/*.(js|jsx)",
-      "src/services/**/*.js"
+      "src/services/**/*.js",
+      "!src/tests/**/*.js"
     ]
   }
 }
