+++ src/pages/Usuarios/Cadastro/Cadastro.test.js
@@ -1,32 +1,22 @@
 import React from 'react';
 import { act, fireEvent, render, screen } from '@testing-library/react';
 import { Route, Router } from 'react-router-dom';
-import userEvent from '@testing-library/user-event';
 import { createMemoryHistory } from 'history';
 import AuthService from '../../../services/Auth/AuthService';
 import { Cadastro } from '../index';
+import { changeInput, historyMock } from '../../../tests/testing';
 
 describe('Cadastro', () => {
-  const changeInput = async (id, value) => {
-    const input = screen.getByTestId(id).querySelector('input');
-    userEvent.type(input, value);
-    fireEvent.blur(input);
-  };
-
   const history = createMemoryHistory();
   history.push('/cadastro');
 
   let pushSpy;
 
   beforeEach(() => {
-    jest.mock('react-router-dom', () => ({
-      ...jest.requireActual('react-router-dom'),
-      useHistory: () => ({
-        push: jest.fn(),
-      }),
-    }));
+    historyMock();
     pushSpy = jest.spyOn(history, 'push');
   });
+
   beforeEach(() => {
     render(
       <Router history={history}>
@@ -37,32 +27,32 @@ describe('Cadastro', () => {
     );
   });
 
-  it('Deve instanciar o componente', async () => {
+  it('Deve instanciar o componente na tela de Cadastro', async () => {
     expect(await screen.findByTestId('cadastroForm')).toBeInTheDocument();
   });
 
-  it('Deve mostrar erro ao colocar um e-mail inválido', async () => {
+  it('Deve mostrar erro ao colocar um e-mail inválido na tela de Cadastro', async () => {
     await changeInput('inputEmail', 'teste Inválido');
     expect(
       await screen.getByText('E-mail informado inválido.')
     ).toBeInTheDocument();
   });
 
-  it('Deve mostrar erro ao colocar uma senha inválida', async () => {
+  it('Deve mostrar erro ao colocar uma senha inválida na tela de Cadastro', async () => {
     await changeInput('inputSenha', 'aa');
     expect(
       await screen.getByText('A senha deve possuir ao menos 3 caracteres.')
     ).toBeInTheDocument();
   });
 
-  it('Deve mostrar erro ao colocar a confirmação de senha inválida', async () => {
+  it('Deve mostrar erro ao colocar a confirmação de senha inválida na tela de Cadastro', async () => {
     await changeInput('inputConfirmacaoSenha', 'aa');
     expect(
       await screen.getByText('A senha deve possuir ao menos 3 caracteres.')
     ).toBeInTheDocument();
   });
 
-  it('Deve mostrar erro ao colocar senhas diferentes', async () => {
+  it('Deve mostrar erro ao colocar senhas diferentes na tela de Cadastro', async () => {
     await act(async () => {
       await changeInput('inputSenha', 'aaa');
       await changeInput('inputConfirmacaoSenha', 'aab');
@@ -72,7 +62,7 @@ describe('Cadastro', () => {
     ).toBeInTheDocument();
   });
 
-  it('Deve habilitar o botão quando o form estiver preenchido corretamente', async () => {
+  it('Deve habilitar o botão quando o form estiver preenchido corretamente na tela de Cadastro', async () => {
     await changeInput('inputEmail', 'teste@valido.com');
     await changeInput('inputSenha', 's3nh4v4l1d4#');
     await changeInput('inputConfirmacaoSenha', 's3nh4v4l1d4#');
@@ -81,7 +71,7 @@ describe('Cadastro', () => {
     ).toBeEnabled();
   });
 
-  it('Deve enviar o Formulário com sucesso', async () => {
+  it('Deve enviar o Formulário com sucesso na tela de Cadastro', async () => {
     jest.spyOn(AuthService, 'cadastrar').mockResolvedValue({
       id: 1,
       email: 'teste@teste.com',
@@ -97,7 +87,7 @@ describe('Cadastro', () => {
     expect(pushSpy).toHaveBeenCalledWith('/usuarios');
   });
 
-  it('Deve dar erro ao enviar o formulário', async () => {
+  it('Deve dar erro ao enviar o formulário na tela de Cadastro', async () => {
     jest.setTimeout(7000);
     jest.spyOn(AuthService, 'cadastrar').mockRejectedValue({
       status: 409,
@@ -115,7 +105,7 @@ describe('Cadastro', () => {
     ).toBeInTheDocument();
   });
 
-  it('Deve voltar para /usuarios ao clicar no botão de voltar', async () => {
+  it('Deve voltar para /usuarios ao clicar no botão de voltar na tela de Cadastro', async () => {
     const button = screen.getByTestId('voltarButton');
     fireEvent.click(button);
